+++ mary/integrations/__init__.py
@@ -5,3 +5,6 @@ __all__=["TwitchPolicy","twitch_policy_from_environment","ObsPolicy","obs_policy
 from .twitch_eventsub import TwitchEventSubConfig, normalize_chat_notification, session_from_welcome
 from .obs_client import ObsHello, authentication_string, identify_message, normalize_obs_event, parse_hello
 __all__ += ["TwitchEventSubConfig","normalize_chat_notification","session_from_welcome","ObsHello","authentication_string","identify_message","normalize_obs_event","parse_hello"]
+
+from .twitch_runtime import EventSubSessionPhase, EventSubControlAction, TwitchEventSubSession, TwitchChatSend, TwitchChatOutbox
+__all__ += ["EventSubSessionPhase","EventSubControlAction","TwitchEventSubSession","TwitchChatSend","TwitchChatOutbox"]